Converted all status applications to Integers
Some grabs may be messed up do to the data conversion.
Most hitstun values are veeeery low now

+++ Assets/Scripts/UI/PlayerCard.cs
@@ -75,7 +75,7 @@ public class PlayerCard : MonoBehaviour
         }
     }
 
-    public GameObject addStatusBar(PlayerStatusEffect statusEffect,int icon, float duration, PlayerStatus status)
+    public GameObject addStatusBar(PlayerStatusEffect statusEffect,int icon, int duration, PlayerStatus status)
     {
 
         GameObject newBar = Instantiate(bar, new Vector3(0,0), Quaternion.identity);
